Use `Int64` instead of `size_t` for sizes in `IMemoryPool`

+++ arccore/src/common/arccore/common/IMemoryPool.h
@@ -48,10 +48,10 @@ class ARCCORE_COMMON_EXPORT IMemoryPool
   virtual void freeCachedMemory() = 0;
 
   //! Taille totale (en octet) allouée dans le pool mémoire
-  virtual size_t totalAllocated() const = 0;
+  virtual Int64 totalAllocated() const = 0;
 
   //! Taille totale (en octet) dans le cache
-  virtual size_t totalCached() const = 0;
+  virtual Int64 totalCached() const = 0;
 };
 
 /*---------------------------------------------------------------------------*/
